user_response_recorder/src/user_response_recorder/requirements.txt
requests
python-dotenv
slack-sdk
urllib3<2